requirements.txt
langchain>=0.0.267
openai>=0.27.8
PyGithub>=1.58.2
python-telegram-bot>=13.15
python-dotenv>=1.0.0
schedule>=1.2.0
pyyaml>=6.0
requests>=2.31.0